Add methods for parsing reviews and users
Previously all parsing was left for downstream processes. This has been changed such that
as reviews are collected the information will be cleaned so that it can be directly
upserted into a collection. This has an advantage for the organization of the users, as
well as making the reviews better structured for downstream applications of model building.